+++ src/constants.ts
@@ -4,6 +4,7 @@ export class Constants {
     public static Context: string = "CharacterGuidelinesActive";
     public static ExtensionName: string = "character-guidelines";
     public static ValidSection: string = "valid-section";
+    public static FillPaddingWithCommand: string = "character-guidelines.fill-padding-with";
     public static ChangeNewLineCountValueCommand: string = "character-guidelines.change-new-line-count-value";
     public static ValidSectionDefault: vscode.DecorationRenderOptions = {
         color: 'green',