+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 
-All notable changes to the "character-guildelines" extension will be documented in this file.
+All notable changes to the "character-guidelines" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
 - Initial release
+
+## [0.1.0]
+
+-   Added the [Character Guidelines: Fill Padding With](#character-guidelines.fill-padding-with) command to populate with real characters the allowed positions